This is a broken commit.
I'm leaving it in the history just because at some point I want to
study how much work in terms of debugging it implies a major
refactoring as the one I'm terminating.

The methods of the original entities.pilot.Pilot class and
the helper classes such as entitities.pilot.Lander have been
redistributed in the newly created pilot package.